+++ CHANGELOG.md
@@ -17,6 +17,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ---
 
+## [1.0.7] - 2026-02-18
+
+### Added
+- Add llm folder in agent to make different llm to locate
+- Add Ollama LLM as default llm
+- Create OllamaLLM class 
+
+### Changed
+- (add changes here)
+
+### Fixed
+- (add fixes here)
+
+---
+
 ## [1.0.6] - 2026-02-18
 
 ### Added